// src/generated/types.ts
export type BidiClass =
  | "arabic_number"
  | "boundary_neutral"
  | "common_number_separator"
  | "european_number"
  | "european_number_separator"
  | "european_number_terminator"
  | "first_strong_isolate"
  | "left_to_right"
  | "left_to_right_embedding"
  | "left_to_right_isolate"
  | "left_to_right_override"
  | "nonspacing_mark"
  | "other_neutrals"
  | "paragraph_separator"
  | "pop_directional_format"
  | "pop_directional_isolate"
  | "right_to_left"
  | "right_to_left_arabic"
  | "right_to_left_embedding"
  | "right_to_left_isolate"
  | "right_to_left_override"
  | "segment_separator"
  | "whitespace";
// ...
export type DecompositionType =
  | "canonical"
  | "circle"
  | "compat"
  | "default"
  | "final"
  | "font"
  | "fraction"
  | "initial"
  | "isolated"
  | "medial"
  | "narrow"
  | "noBreak"
  | "small"
  | "square"
  | "sub"
  | "super"
  | "vertical"
  | "wide";

export type EastAsianWidth =
  | "ambiguous"
  | "fullwidth"
  | "halfwidth"
  | "narrow"
  | "neutral"
  | "wide";

export type GeneralCategory =
  | "letter_lowercase"
  | "letter_modifier"
  | "letter_other"
  | "letter_titlecase"
  | "letter_uppercase"
  | "mark_enclosing"
  | "mark_nonspacing"
  | "mark_spacing_combining"
  | "number_decimal_digit"
  | "number_letter"
  | "number_other"
  | "other_control"
  | "other_format"
  | "other_not_assigned"
  | "other_private_use"
  | "other_surrogate"
  | "punctuation_close"
  | "punctuation_connector"
  | "punctuation_dash"
  | "punctuation_final_quote"
  | "punctuation_initial_quote"
  | "punctuation_open"
  | "punctuation_other"
  | "separator_line"
  | "separator_paragraph"
  | "separator_space"
  | "symbol_currency"
  | "symbol_math"
  | "symbol_modifier"
  | "symbol_other";

export type GraphemeBreakProperty =
  | "control"
  | "cr"
  | "emoji_modifier"
  | "emoji_modifier_base"
  | "extended_pictographic"
  | "indic_conjunct_break_consonant"
  | "indic_conjunct_break_extend"
  | "indic_conjunct_break_linker"
  | "l"
  | "lf"
  | "lv"
  | "lvt"
  | "other"
  | "prepend"
  | "regional_indicator"
  | "spacing_mark"
  | "t"
  | "v"
  | "zwj"
  | "zwnj";

export type GraphemeBreakNoControlProperty =
  | "emoji_modifier"
  | "emoji_modifier_base"
  | "extended_pictographic"
  | "indic_conjunct_break_consonant"
  | "indic_conjunct_break_extend"
  | "indic_conjunct_break_linker"
  | "l"
  | "lv"
  | "lvt"
  | "other"
  | "prepend"
  | "regional_indicator"
  | "spacing_mark"
  | "t"
  | "v"
  | "zwj"
  | "zwnj";

export type IndicConjunctBreak = "consonant" | "extend" | "linker" | "none";

export type IndicPositionalCategory =
  | "bottom"
  | "bottom_and_left"
  | "bottom_and_right"
  | "left"
  | "left_and_right"
  | "not_applicable"
  | "overstruck"
  | "right"
  | "top"
  | "top_and_bottom"
  | "top_and_bottom_and_left"
  | "top_and_bottom_and_right"
  | "top_and_left"
  | "top_and_left_and_right"
  | "top_and_right"
  | "visual_order_left";

export type IndicSyllabicCategory =
  | "avagraha"
  | "bindu"
  | "brahmi_joining_number"
  | "cantillation_mark"
  | "consonant"
  | "consonant_dead"
  | "consonant_final"
  | "consonant_head_letter"
  | "consonant_initial_postfixed"
  | "consonant_killer"
  | "consonant_medial"
  | "consonant_placeholder"
  | "consonant_preceding_repha"
  | "consonant_prefixed"
  | "consonant_subjoined"
  | "consonant_succeeding_repha"
  | "consonant_with_stacker"
  | "gemination_mark"
  | "invisible_stacker"
  | "joiner"
  | "modifying_letter"
  | "non_joiner"
  | "nukta"
  | "number"
  | "number_joiner"
  | "other"
  | "pure_killer"
  | "register_shifter"
  | "reordering_killer"
  | "syllable_modifier"
  | "tone_letter"
  | "tone_mark"
  | "virama"
  | "visarga"
  | "vowel"
  | "vowel_dependent"
  | "vowel_independent";

export type JoiningGroup =
  | "african_feh"
  | "african_noon"
  | "african_qaf"
  | "ain"
  | "alaph"
  | "alef"
  | "beh"
  | "beth"
  | "burushaski_yeh_barree"
  | "dal"
  | "dalath_rish"
  | "e"
  | "farsi_yeh"
  | "fe"
  | "feh"
  | "final_semkath"
  | "gaf"
  | "gamal"
  | "hah"
  | "hanifi_rohingya_kinna_ya"
  | "hanifi_rohingya_pa"
  | "he"
  | "heh"
  | "heh_goal"
  | "heth"
  | "kaf"
  | "kaph"
  | "kashmiri_yeh"
  | "khaph"
  | "knotted_heh"
  | "lam"
  | "lamadh"
  | "malayalam_bha"
  | "malayalam_ja"
  | "malayalam_lla"
  | "malayalam_llla"
  | "malayalam_nga"
  | "malayalam_nna"
  | "malayalam_nnna"
  | "malayalam_nya"
  | "malayalam_ra"
  | "malayalam_ssa"
  | "malayalam_tta"
  | "manichaean_aleph"
  | "manichaean_ayin"
  | "manichaean_beth"
  | "manichaean_daleth"
  | "manichaean_dhamedh"
  | "manichaean_five"
  | "manichaean_gimel"
  | "manichaean_heth"
  | "manichaean_hundred"
  | "manichaean_kaph"
  | "manichaean_lamedh"
  | "manichaean_mem"
  | "manichaean_nun"
  | "manichaean_one"
  | "manichaean_pe"
  | "manichaean_qoph"
  | "manichaean_resh"
  | "manichaean_sadhe"
  | "manichaean_samekh"
  | "manichaean_taw"
  | "manichaean_ten"
  | "manichaean_teth"
  | "manichaean_thamedh"
  | "manichaean_twenty"
  | "manichaean_waw"
  | "manichaean_yodh"
  | "manichaean_zayin"
  | "meem"
  | "mim"
  | "no_joining_group"
  | "noon"
  | "nun"
  | "nya"
  | "pe"
  | "qaf"
  | "qaph"
  | "reh"
  | "reversed_pe"
  | "rohingya_yeh"
  | "sad"
  | "sadhe"
  | "seen"
  | "semkath"
  | "shin"
  | "straight_waw"
  | "swash_kaf"
  | "syriac_waw"
  | "tah"
  | "taw"
  | "teh_marbuta"
  | "teh_marbuta_goal"
  | "teth"
  | "thin_noon"
  | "thin_yeh"
  | "vertical_tail"
  | "waw"
  | "yeh"
  | "yeh_barree"
  | "yeh_with_tail"
  | "yudh"
  | "yudh_he"
  | "zain"
  | "zhain";

export type JoiningType =
  | "dual_joining"
  | "join_causing"
  | "left_joining"
  | "non_joining"
  | "right_joining"
  | "transparent";

export type NumericType = "decimal" | "digit" | "none" | "numeric";

export type OriginalGraphemeBreak =
  | "control"
  | "cr"
  | "extend"
  | "l"
  | "lf"
  | "lv"
  | "lvt"
  | "other"
  | "prepend"
  | "regional_indicator"
  | "spacing_mark"
  | "t"
  | "v"
  | "zwj";

export type Script =
  | "adlam"
  | "ahom"
  | "anatolian_hieroglyphs"
  | "arabic"
  | "armenian"
  | "avestan"
  | "balinese"
  | "bamum"
  | "bassa_vah"
  | "batak"
  | "bengali"
  | "beria_erfe"
  | "bhaiksuki"
  | "bopomofo"
  | "brahmi"
  | "braille"
  | "buginese"
  | "buhid"
  | "canadian_aboriginal"
  | "carian"
  | "caucasian_albanian"
  | "chakma"
  | "cham"
  | "cherokee"
  | "chorasmian"
  | "common"
  | "coptic"
  | "cuneiform"
  | "cypriot"
  | "cypro_minoan"
  | "cyrillic"
  | "deseret"
  | "devanagari"
  | "dives_akuru"
  | "dogra"
  | "duployan"
  | "egyptian_hieroglyphs"
  | "elbasan"
  | "elymaic"
  | "ethiopic"
  | "garay"
  | "georgian"
  | "glagolitic"
  | "gothic"
  | "grantha"
  | "greek"
  | "gujarati"
  | "gunjala_gondi"
  | "gurmukhi"
  | "gurung_khema"
  | "han"
  | "hangul"
  | "hanifi_rohingya"
  | "hanunoo"
  | "hatran"
  | "hebrew"
  | "hiragana"
  | "imperial_aramaic"
  | "inherited"
  | "inscriptional_pahlavi"
  | "inscriptional_parthian"
  | "javanese"
  | "kaithi"
  | "kannada"
  | "katakana"
  | "kawi"
  | "kayah_li"
  | "kharoshthi"
  | "khitan_small_script"
  | "khmer"
  | "khojki"
  | "khudawadi"
  | "kirat_rai"
  | "lao"
  | "latin"
  | "lepcha"
  | "limbu"
  | "linear_a"
  | "linear_b"
  | "lisu"
  | "lycian"
  | "lydian"
  | "mahajani"
  | "makasar"
  | "malayalam"
  | "mandaic"
  | "manichaean"
  | "marchen"
  | "masaram_gondi"
  | "medefaidrin"
  | "meetei_mayek"
  | "mende_kikakui"
  | "meroitic_cursive"
  | "meroitic_hieroglyphs"
  | "miao"
  | "modi"
  | "mongolian"
  | "mro"
  | "multani"
  | "myanmar"
  | "nabataean"
  | "nag_mundari"
  | "nandinagari"
  | "new_tai_lue"
  | "newa"
  | "nko"
  | "nushu"
  | "nyiakeng_puachue_hmong"
  | "ogham"
  | "ol_chiki"
  | "ol_onal"
  | "old_hungarian"
  | "old_italic"
  | "old_north_arabian"
  | "old_permic"
  | "old_persian"
  | "old_sogdian"
  | "old_south_arabian"
  | "old_turkic"
  | "old_uyghur"
  | "oriya"
  | "osage"
  | "osmanya"
  | "pahawh_hmong"
  | "palmyrene"
  | "pau_cin_hau"
  | "phags_pa"
  | "phoenician"
  | "psalter_pahlavi"
  | "rejang"
  | "runic"
  | "samaritan"
  | "saurashtra"
  | "sharada"
  | "shavian"
  | "siddham"
  | "sidetic"
  | "sign_writing"
  | "sinhala"
  | "sogdian"
  | "sora_sompeng"
  | "soyombo"
  | "sundanese"
  | "sunuwar"
  | "syloti_nagri"
  | "syriac"
  | "tagalog"
  | "tagbanwa"
  | "tai_le"
  | "tai_tham"
  | "tai_viet"
  | "tai_yo"
  | "takri"
  | "tamil"
  | "tangsa"
  | "tangut"
  | "telugu"
  | "thaana"
  | "thai"
  | "tibetan"
  | "tifinagh"
  | "tirhuta"
  | "todhri"
  | "tolong_siki"
  | "toto"
  | "tulu_tigalari"
  | "ugaritic"
  | "unknown"
  | "vai"
  | "vithkuqi"
  | "wancho"
  | "warang_citi"
  | "yezidi"
  | "yi"
  | "zanabazar_square";

export type SpecialCasingCondition =
  | "after_i"
  | "az"
  | "final_sigma"
  | "lt"
  | "more_above"
  | "not_before_dot";